add css shopping list
and rename files
