+++ src/css/shoppingList.css
@@ -1,0 +1,280 @@
+.main-content{
+    padding-bottom: 40px;
+    justify-content: center;
+    display: flex;
+}
+
+
+.shopping-title {
+
+    font-weight: 700;
+    font-size: 32px;
+    line-height: calc(38 / 32);
+
+    letter-spacing: -0.04em;
+
+    color: #111;
+
+    margin-bottom: 40px;
+ 
+}
+
+.shopping-title-span {
+    color: #4f2ee8;
+}
+/* // EMPTY */
+.shop-card-empty-text {
+    font-size: 14px;
+    line-height: calc(18 / 14);
+
+    text-align: center;
+    letter-spacing: -0.02em;
+    color: #0009;}
+
+.shop-card-empty {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        padding: 0px;
+        gap: 14px;
+
+        margin-top: 120px;
+        margin-left: auto;
+        margin-right: auto;
+
+        width: 265px;
+        height: 248px;}
+    
+    
+.shop-card-empty-picture {
+    width: 265px;
+    height: 198px;
+}
+
+/*============================================================================================== */
+.shopping-list{
+
+}
+.shopping__card {
+    width: 100%;
+    
+        display: grid;
+        /* grid-template-columns: 98px 1fr;
+        grid-template-rows: 60px 79px 34px 1fr; */
+        grid-column-gap: 20px;
+        grid-row-gap: 0px;
+    
+        position: relative;
+    
+        background-color: white;
+        width: 336px;
+        height: auto;
+        margin-left: auto;
+        margin-right: auto;
+        margin-bottom: 20px;
+        padding: 14px;
+        padding-bottom: 20px;
+    
+        border: 3px solid rgba(79, 46, 232, 0.4);
+        border-radius: 16px;
+    
+        
+}
+.shopping-card:hover,
+.shopping-card:focus {
+    border: 3px solid #eac645;
+}
+.about-img {grid-area: 1 / 1 / span 3 / 2;
+width: 100px;
+    height: 142px;}
+
+.shopping-card-img {
+    width: 100px;
+        height: 142px;
+
+    border-radius: 8px;
+}
+
+.about-title {grid-area: 1 / 2 / 1 / 4;
+    height: 36px;
+}
+
+.shopping-card-title {margin-bottom: 4px;
+
+    font-weight: 700;
+    font-size: 16px;
+    line-height: calc(18 / 16);
+
+    letter-spacing: -0.04em;
+    text-transform: uppercase;
+
+    color: #111111;}
+
+.shopping-card-category {
+    font-family: 'DM Sans';
+    font-size: 12px;
+    line-height: calc(14 / 12);
+
+    letter-spacing: 0.03em;
+
+    color: #B4AFAF;;}
+
+
+.about-description{
+    grid-area: 2 / 2 / 2 / 4;
+    margin-bottom: 14px;
+}
+.shopping-card-description {
+    overflow: auto;
+    margin-top: 8px;
+
+    font-family: 'DM Sans';
+    font-size: 15px;
+
+    letter-spacing: -0.04em;
+
+    color: var(--text-color);
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 1.29;
+    
+    color: #111111;
+    max-height: 68px;}
+
+
+.about-author {
+    grid-area: 3 / 2 / 3 / 2;
+}
+
+.shopping-card-author {
+    margin-top: 4px;
+    
+        font-style: italic;
+        font-size: 12px;
+        line-height: calc(14 / 12);
+    
+        letter-spacing: -0.04em;
+    
+        color: #B4AFAF;;
+    }
+
+.shoplist-url {
+grid-area: 3 / 3 / 3 / 4;
+}
+
+.shoplist-url-list {
+    display: flex;
+    gap: 14px;
+    align-items: center;
+    mix-blend-mode: luminosity;
+}
+
+.shoplist-url-item {
+
+}
+
+.shoplist-url-link {}
+
+.modal-shop-img {}
+
+.shopping-shopimg {}
+
+.amazon {}
+
+.apple {}
+
+.book-shop {}
+
+svg{max-width: 100%;
+    height: auto;
+    display: block;}
+
+.shopping-card-btn {display: flex;
+    align-items: center;
+    justify-content: center;
+
+    position: absolute;
+    top: 14px;
+    right: 14px;
+
+    width: 28px;
+    height: 28px;
+    border-radius: 50%;
+    border-color: transparent;
+    background-color: #4f2ee8;
+
+    transition: background 0.2s ease-out;}
+
+.icon-trash {}
+.shopping__card-btn:hover,
+.shopping__card-btn:hover {
+     transform: scale(1.1);
+    background: #eac645;
+}
+
+
+
+
+@media screen and (max-width: 767.99px) {
+    .main-content{
+    display: flex;
+    justify-content: center;}
+
+}
+ @media screen and (min-width: 768px) {
+     .main-content{ flex-grow: 1;
+     padding-bottom: 60px;}
+
+     .shopping-title{
+        font-size: 48px;
+        line-height: calc(52 / 48);
+     }
+     .shop-card-empty-text{
+        font-size: 18px;
+            line-height: calc(24 / 18);
+            margin-left: 0px;
+            margin-right: 0px;
+     }
+     
+     .shopping__card{
+        grid-template-columns: 116px repeat(2, 1fr);
+        grid-template-rows: 50px 85px 30px;
+        grid-column-gap: 24px;
+    
+        width: 100%;
+        max-height: 213px;
+        padding: 24px;}
+
+    .about-author{
+        max-width: 97px;
+        grid-area: 3 / 2 / 4 / 3;
+    
+        display: flex;
+        align-items: end;
+
+    }
+
+    .shopping-card-description{
+        grid-area: 2 / 2 / 3 / 4;
+    }
+
+    .shopping-card-title{grid-area: 1/2/2/4;
+
+    }
+
+    .shoplist-url-list{
+        grid-area: 3 / 3 / 4 / 4;
+        
+            display: flex;
+            justify-content: end;
+            align-items: end;
+    }
+    .shopping-card-img{
+        width: 116px;
+            height: 165px;
+    }
+ }
+
+ @media screen and (min-width: 1440px) {
+     .main-content{ flex-grow: 1;}
+ }
