+++ src/css/styles.css
@@ -23,3 +23,5 @@
 @import url(/src/css/loaders.css);
 
 @import url(/src/css/header.css);
+/* ShoppingList */
+@import url(/src/css/shoppingList.css);
