+++ src/main.js
@@ -6,7 +6,7 @@ import './js/support-asside/support-functionality'
 import './js/homePageBookRender'
 import './js/home/allCategoriesList'
 import './js/ShoppingList'
-import './js/shopping-list'
+import './js/shoppingListContent'
 // import {BooksAPIService} from './js/booksAPIService'
 import './js/home/booksByCategories'
