+++ src/partials/shopping-list-content.html
@@ -1,5 +1,5 @@
 <section class="shopping-list-content">
-  <div class="shopping-container">
+  <div class="shopping-container container">
     <h2 class="shopping-title">
       Shopping <span class="shopping-title-span">List</span>
     </h2>
